oauth_login/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Django==5.1.3
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
idna==3.10
jwcrypto==1.5.6
mysqlclient==2.2.6
oauthlib==3.2.2
pycparser==2.22
requests==2.32.3
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
